requirements.txt
gymnasium[box2d]>=1.0.0
stable-baselines3>=2.3.0
torch
torchvision
numpy
scipy
pandas
matplotlib
pillow
